Rename signer-with-address.ts to signers.ts
