+++ packages/hardhat-ethers/src/internal/helpers.ts
@@ -6,7 +6,7 @@ import {
   NetworkConfig,
 } from "hardhat/types";
 
-import type { SignerWithAddress } from "../signer-with-address";
+import type { SignerWithAddress } from "../signers";
 import type { FactoryOptions, Libraries } from "../types";
 
 interface Link {
@@ -21,7 +21,7 @@ export async function getSigners(
   hre: HardhatRuntimeEnvironment
 ): Promise<SignerWithAddress[]> {
   const { SignerWithAddress: SignerWithAddressImpl } = await import(
-    "../signer-with-address"
+    "../signers"
   );
 
   const accounts = await hre.ethers.provider.listAccounts();
